--- source/corvusoft/restbed/detail/common_impl.hpp
@@ -1,149 +1,0 @@
-/*
- * Copyright 2013-2016, Corvusoft Ltd, All Rights Reserved.
- */
-
-#ifndef _RESTBED_DETAIL_COMMON_IMPL_H
-#define _RESTBED_DETAIL_COMMON_IMPL_H 1
-
-//System Includes
-#include <string>
-#include <utility>
-#include <sstream>
-#include <algorithm>
-#include <type_traits>
-
-//Project Includes
-#include <corvusoft/restbed/string.hpp>
-
-//External Includes
-
-//System Namespaces
-
-//Project Namespaces
-
-//External Namespaces
-
-namespace restbed
-{
-    //Forward Declarations
-    
-    namespace detail
-    {
-        //Forward Declarations
-        
-        class CommonImpl
-        {
-            public:
-                //Friends
-                
-                //Definitions
-                
-                //Constructors
-                
-                //Functionality
-                template< typename Type >
-                static Type parse_parameter( const std::string& value, const Type default_value )
-                {
-                    std::istringstream stream( value );
-                    
-                    Type parameter;
-                    stream >> parameter;
-                    
-                    if ( stream.fail( ) )
-                    {
-                        return default_value;
-                    }
-                    
-                    return parameter;
-                }
-                
-                template< typename Type >
-                static bool has_parameter( const std::string& name, const Type& parameters )
-                {
-                    const auto key = String::lowercase( name );
-                    const auto iterator = std::find_if( parameters.begin( ), parameters.end( ), [ &key ]( const std::pair< std::string, std::string >& value )
-                    {
-                        return ( key == String::lowercase( value.first ) );
-                    } );
-                    
-                    return iterator not_eq parameters.end( );
-                }
-                
-                static std::string transform( const std::string& value, const std::function< std::string ( const std::string& ) >& transform )
-                {
-                    return ( transform == nullptr ) ? value : transform( value );
-                }
-                
-                //Getters
-                template< typename Type >
-                static Type get_parameters( const std::string& name, const Type& parameters )
-                {
-                    if ( name.empty( ) )
-                    {
-                        return parameters;
-                    }
-                    
-                    const auto key = String::lowercase( name );
-                    Type results;
-                    
-                    for ( const auto& parameter : parameters )
-                    {
-                        if ( key == String::lowercase( parameter.first ) )
-                        {
-                            results.insert( parameter );
-                        }
-                    }
-                    
-                    return results;
-                }
-                
-                //Setters
-                
-                //Operators
-                
-                //Properties
-                
-            protected:
-                //Friends
-                
-                //Definitions
-                
-                //Constructors
-                
-                //Functionality
-                
-                //Getters
-                
-                //Setters
-                
-                //Operators
-                
-                //Properties
-                
-            private:
-                //Friends
-                
-                //Definitions
-                
-                //Constructors
-                CommonImpl( void ) = delete;
-                
-                CommonImpl( const CommonImpl& original ) = delete;
-                
-                virtual ~CommonImpl( void ) = delete;
-                
-                //Functionality
-                
-                //Getters
-                
-                //Setters
-                
-                //Operators
-                CommonImpl& operator =( const CommonImpl& value ) = delete;
-                
-                //Properties
-        };
-    }
-}
-
-#endif  /* _RESTBED_DETAIL_COMMON_IMPL_H */
